Productivity Client, Tasks module, created tasks api endpoints
- Separate tasks api endpoints from the main api file
