+++ src/features/Tasks/hooks/useCreateTask.js
@@ -1,7 +1,7 @@
 import { useCheckAuth } from "../../../context/AuthenticationContext";
 import { useDispatch } from "react-redux";
 import { createTask } from "../tasksSlice";
-import { useCreateTaskMutation } from "../../../api/api";
+import { useCreateTaskMutation } from "../../../api/tasks/tasks";
 import { useState } from "react";
 import moment from "moment";
